+++ scheduling/flow/cluster/arc.go
@@ -14,7 +14,7 @@
 
 // Represents an arc in the scheduling flow graph.
 // C++ file: https://github.com/camsas/firmament/blob/master/src/scheduling/flow/flow_graph_arc.h
-package graph
+package cluster
 
 //Enum for flow arc type
 type FlowArcType int
