+++ scheduling/flow/cluster/graph.go
@@ -15,7 +15,7 @@
 // Represents the scheduling flow graph.
 // C++ file: https://github.com/camsas/firmament/blob/master/src/scheduling/flow/flow_graph.h
 
-package graph
+package cluster
 
 import (
 	"log"
